pyproject.toml
[project]
name = "sci-markdown"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "mdit-py-plugins>=0.4.2",
    "markdown-it-py>=3.0.0",
    "watchdog>=5.0.2",
    "numpy>=2.1.1",
    "matplotlib>=3.9.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"sci-markdown" = "sci_markdown:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/sci_markdown"]
